+++ main_gpu.cu
@@ -5,24 +5,24 @@
 #include "layers/dense.h"
 #include "layers/pooling.h"
 
-void test_dense(double *input) {
+void test_dense(float *input) {
     Dense d(6, 10);
     d.dump();
 
-    double *doutput = d.forward(input);
-    double *output = (double*)malloc(sizeof(double) * 10);
-    cudaMemcpy(output, doutput, sizeof(double) * 10, cudaMemcpyDeviceToHost);
+    float *doutput = d.forward(input);
+    float *output = (float*)malloc(sizeof(float) * 10);
+    cudaMemcpy(output, doutput, sizeof(float) * 10, cudaMemcpyDeviceToHost);
 
     printf("\nOutput: ");
     for (int i=0; i<10; i++) printf("%9.6lf ", output[i]);
     printf("\n");
 
-    double *loss;
-    cudaMalloc(&loss, sizeof(double) * 10);
+    float *loss;
+    cudaMalloc(&loss, sizeof(float) * 10);
     for (int i=0; i<10; i++) output[i] = 2;
     printf("\nLoss: ");
     for (int i=0; i<10; i++) printf("%9.6lf ", output[i]);
-    cudaMemcpy(loss, output, sizeof(double) * 10, cudaMemcpyHostToDevice);
+    cudaMemcpy(loss, output, sizeof(float) * 10, cudaMemcpyHostToDevice);
 
     d.backward(loss, 0.1);
     d.dump();
@@ -31,13 +31,13 @@ void test_dense(double *input) {
     cudaFree(loss);
 }
 
-void test_pooling(double *input, int c, int h, int w, int s) {
+void test_pooling(float *input, int c, int h, int w, int s) {
     Pooling p(c, h, w, s);
     p.dump();
 
-    double *doutput = p.forward(input);
-    double *output = (double*)malloc(sizeof(double) * c*h/s*w/s);
-    cudaMemcpy(output, doutput, sizeof(double) * c*h/s*w/s, cudaMemcpyDeviceToHost);
+    float *doutput = p.forward(input);
+    float *output = (float*)malloc(sizeof(float) * c*h/s*w/s);
+    cudaMemcpy(output, doutput, sizeof(float) * c*h/s*w/s, cudaMemcpyDeviceToHost);
 
     printf("Output:\n");
     for (int i=0; i<c; i++) {
@@ -48,8 +48,8 @@ void test_pooling(double *input, int c, int h, int w, int s) {
         printf("\n");
     }
 
-    double *loss;
-    cudaMalloc(&loss, sizeof(double) * c*h/s*w/s);
+    float *loss;
+    cudaMalloc(&loss, sizeof(float) * c*h/s*w/s);
     for (int i=0; i<c*h/s*w/s; i++) output[i] = 99;
     printf("\nLoss:\n");
     for (int i=0; i<c; i++) {
@@ -59,11 +59,11 @@ void test_pooling(double *input, int c, int h, int w, int s) {
         }
         printf("\n");
     }
-    cudaMemcpy(loss, output, sizeof(double) * c*h/s*w/s, cudaMemcpyHostToDevice);
+    cudaMemcpy(loss, output, sizeof(float) * c*h/s*w/s, cudaMemcpyHostToDevice);
 
-    double *dd = p.backward(loss, 0.1);
-    double *d = (double*)malloc(sizeof(double)*c*h*w);
-    cudaMemcpy(d, dd, sizeof(double)*c*h*w, cudaMemcpyDeviceToHost);
+    float *dd = p.backward(loss, 0.1);
+    float *d = (float*)malloc(sizeof(float)*c*h*w);
+    cudaMemcpy(d, dd, sizeof(float)*c*h*w, cudaMemcpyDeviceToHost);
 
     printf("Output loss:\n");
     for (int i=0; i<c; i++) {
@@ -79,15 +79,15 @@ void test_pooling(double *input, int c, int h, int w, int s) {
     cudaFree(loss);
 }
 
-void test_conv(double *input, int c, int h, int w, int oc, int k, int s, int p) {
+void test_conv(float *input, int c, int h, int w, int oc, int k, int s, int p) {
     Conv conv(c, h, w, oc, k, s, p);
     conv.dump();
 
     int oh = (h+2*p-k)/s+1;
     int ow = (w+2*p-k)/s+1;
-    double *doutput = conv.forward(input);
-    double *output = (double*)malloc(sizeof(double) * oc*oh*ow);
-    cudaMemcpy(output, doutput, sizeof(double) * oc*oh*ow, cudaMemcpyDeviceToHost);
+    float *doutput = conv.forward(input);
+    float *output = (float*)malloc(sizeof(float) * oc*oh*ow);
+    cudaMemcpy(output, doutput, sizeof(float) * oc*oh*ow, cudaMemcpyDeviceToHost);
 
     printf("Output:\n");
     for (int i=0; i<oc; i++) {
@@ -98,8 +98,8 @@ void test_conv(double *input, int c, int h, int w, int oc, int k, int s, int p) 
         printf("\n");
     }
 
-    double *loss;
-    cudaMalloc(&loss, sizeof(double) * oc*oh*ow);
+    float *loss;
+    cudaMalloc(&loss, sizeof(float) * oc*oh*ow);
     for (int i=0; i<oc*oh*ow; i++) output[i] = 2;
     printf("\nLoss:\n");
     for (int i=0; i<oc; i++) {
@@ -109,12 +109,12 @@ void test_conv(double *input, int c, int h, int w, int oc, int k, int s, int p) 
         }
         printf("\n");
     }
-    cudaMemcpy(loss, output, sizeof(double) * oc*oh*ow, cudaMemcpyHostToDevice);
+    cudaMemcpy(loss, output, sizeof(float) * oc*oh*ow, cudaMemcpyHostToDevice);
 
-    double *dd = conv.backward(loss, 0.1);
+    float *dd = conv.backward(loss, 0.1);
     conv.dump();
-    double *d = (double*)malloc(sizeof(double)*c*h*w);
-    cudaMemcpy(d, dd, sizeof(double)*c*h*w, cudaMemcpyDeviceToHost);
+    float *d = (float*)malloc(sizeof(float)*c*h*w);
+    cudaMemcpy(d, dd, sizeof(float)*c*h*w, cudaMemcpyDeviceToHost);
 
     printf("Output loss:\n");
     for (int i=0; i<c; i++) {
@@ -133,11 +133,11 @@ void test_conv(double *input, int c, int h, int w, int oc, int k, int s, int p) 
 int main() {
     rand_init();
     int c = 1, h = 5, w = 5;
-    double *input = (double*)malloc(sizeof(double) * c*h*w);
+    float *input = (float*)malloc(sizeof(float) * c*h*w);
     for (int i=0; i<c*h*w; i++) input[i] = 1;//randn();
-    double *dinput;
-    cudaMalloc(&dinput, sizeof(double) * c*h*w);
-    cudaMemcpy(dinput, input, sizeof(double) * c*h*w, cudaMemcpyHostToDevice);
+    float *dinput;
+    cudaMalloc(&dinput, sizeof(float) * c*h*w);
+    cudaMemcpy(dinput, input, sizeof(float) * c*h*w, cudaMemcpyHostToDevice);
 
     printf("Input:\n");
     for (int i=0; i<c; i++) {
